Utils cleanup
