+++ Plugin/src/main/java/com/gmail/filoghost/chestcommands/util/StringUtils.java
@@ -14,7 +14,12 @@
  */
 package com.gmail.filoghost.chestcommands.util;
 
-public class StringUtils {
+import java.util.Iterator;
+
+public final class StringUtils {
+
+	private StringUtils() {
+	}
 
 	public static String stripChars(String input, String removed) {
 		if (removed == null || removed.isEmpty()) {
@@ -98,5 +103,24 @@ public class StringUtils {
 	public static boolean isNullOrEmpty(String s) {
 		return s == null || s.isEmpty();
 	}
-	
+
+    public static String join(Iterable<?> iterable, String separator) {
+        StringBuilder builder = new StringBuilder();
+        Iterator<?> iter = iterable.iterator();
+
+        boolean first = true;
+
+        while (iter.hasNext()) {
+            if (first) {
+                first = false;
+            } else {
+                builder.append(separator);
+            }
+
+            builder.append(iter.next());
+        }
+
+        return builder.toString();
+    }
+
 }
